back/pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "faelbochi", email = ""},
]
dependencies = [
    "django>=4.2.3",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.0.0",
    "drf-spectacular>=0.26.3",
    "django-cors-headers>=4.1.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "django-extensions>=3.2.3",
    "python-magic-bin==0.4.14",
]
requires-python = ">=3.11"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"